requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
motor==3.6.0
pydantic==2.9.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
jinja2==3.1.4
gunicorn==22.0.0
python-dotenv==1.0.1
